+++ drools-planner-core/src/main/java/org/drools/planner/core/score/HardAndSoftScore.java
@@ -32,7 +32,7 @@ public interface HardAndSoftScore extends Score<HardAndSoftScore> {
      * The hard score is usually a negative number because most use cases only have negative constraints.
      * @return higher is better, usually negative, 0 if no hard constraints are broken/fulfilled
      */
-    int getHardScore();
+    long getHardScore();
 
     /**
      * The total of the broken negative soft constraints and fulfilled positive soft constraints.
@@ -42,6 +42,6 @@ public interface HardAndSoftScore extends Score<HardAndSoftScore> {
      * In a normal score comparison, the soft score is irrelevant if the 2 scores don't have the same hard score.
      * @return higher is better, usually negative, 0 if no soft constraints are broken/fulfilled
      */
-    int getSoftScore();
+    long getSoftScore();
 
 }
